+++ app/model/entity/flink.go
@@ -5,7 +5,7 @@ type Flink struct {
   Name string `json:"name" form:"name"`
   Img string `json:"img" form:"img"`
   Url string `json:"url" form:"url"`
-  CateId int64 `json:"cate_id" form:"cate_id"`
-  Ordid int64 `json:"ordid" form:"ordid"`
-  Status int64 `json:"status" form:"status"`
+  CateId int16 `json:"cate_id" form:"cate_id"`
+  Ordid int8 `json:"ordid" form:"ordid"`
+  Status int8 `json:"status" form:"status"`
 }
