+++ app/model/entity/flink_cate.go
@@ -3,6 +3,6 @@ package entity
 type FlinkCate struct {
   Model
   Name string `json:"name" form:"name"`
-  Ordid int64 `json:"ordid" form:"ordid"`
-  Status int64 `json:"status" form:"status"`
+  Ordid int8 `json:"ordid" form:"ordid"`
+  Status int8 `json:"status" form:"status"`
 }
