+++ app/model/entity/reward_gif.go
@@ -3,10 +3,10 @@ package entity
 type RewardGif struct {
   Model
   Name string `json:"name" form:"name"`
-  Bit int64 `json:"bit" form:"bit"`
+  Bit int32 `json:"bit" form:"bit"`
   Img string `json:"img" form:"img"`
   Flash string `json:"flash" form:"flash"`
   FlashTime float64 `json:"flash_time" form:"flash_time"`
-  IsVip int64 `json:"is_vip" form:"is_vip"`
-  AdminId int64 `json:"admin_id" form:"admin_id"`
+  IsVip int8 `json:"is_vip" form:"is_vip"`
+  AdminId int32 `json:"admin_id" form:"admin_id"`
 }
