+++ app/model/entity/site_msg.go
@@ -6,7 +6,7 @@ type SiteMsg struct {
   Sid int64 `json:"sid" form:"sid"`
   Title string `json:"title" form:"title"`
   Info string `json:"info" form:"info"`
-  Type int64 `json:"type" form:"type"`
+  Type int8 `json:"type" form:"type"`
   Intro string `json:"intro" form:"intro"`
-  IsRead int64 `json:"is_read" form:"is_read"`
+  IsRead int8 `json:"is_read" form:"is_read"`
 }
