+++ app/model/entity/sms_record.go
@@ -3,7 +3,7 @@ package entity
 type SmsRecord struct {
   Model
   Phone string `json:"phone" form:"phone"`
-  Status int64 `json:"status" form:"status"`
+  Status int8 `json:"status" form:"status"`
   Content string `json:"content" form:"content"`
-  AddIp int64 `json:"add_ip" form:"add_ip"`
+  AddIp int32 `json:"add_ip" form:"add_ip"`
 }
