+++ app/model/entity/sys_notice.go
@@ -5,5 +5,5 @@ type SysNotice struct {
   Title string `json:"title" form:"title"`
   Intro string `json:"intro" form:"intro"`
   Info string `json:"info" form:"info"`
-  Status int64 `json:"status" form:"status"`
+  Status int8 `json:"status" form:"status"`
 }
