+++ app/model/entity/user_address.go
@@ -3,9 +3,9 @@ package entity
 type UserAddress struct {
   Model
   Uid int64 `json:"uid" form:"uid"`
-  Pid int64 `json:"pid" form:"pid"`
-  Cid int64 `json:"cid" form:"cid"`
-  Aid int64 `json:"aid" form:"aid"`
+  Pid int32 `json:"pid" form:"pid"`
+  Cid int32 `json:"cid" form:"cid"`
+  Aid int32 `json:"aid" form:"aid"`
   Province string `json:"province" form:"province"`
   City string `json:"city" form:"city"`
   Area string `json:"area" form:"area"`
@@ -14,5 +14,5 @@ type UserAddress struct {
   Realname string `json:"realname" form:"realname"`
   Phone string `json:"phone" form:"phone"`
   Postalcode string `json:"postalcode" form:"postalcode"`
-  Default int64 `json:"default" form:"default"`
+  Default int8 `json:"default" form:"default"`
 }
