+++ app/model/entity/user_asset_log_type.go
@@ -2,6 +2,6 @@ package entity
 
 type UserAssetLogType struct {
   Model
-  Type int64 `json:"type" form:"type"`
+  Type int16 `json:"type" form:"type"`
   Val string `json:"val" form:"val"`
 }
