+++ app/model/entity/user_demand.go
@@ -4,9 +4,9 @@ type UserDemand struct {
   Model
   Uid int64 `json:"uid" form:"uid"`
   Tid string `json:"tid" form:"tid"`
-  WorkYear int64 `json:"work_year" form:"work_year"`
-  Area int64 `json:"area" form:"area"`
-  Budget int64 `json:"budget" form:"budget"`
+  WorkYear int8 `json:"work_year" form:"work_year"`
+  Area int32 `json:"area" form:"area"`
+  Budget int32 `json:"budget" form:"budget"`
   Demand string `json:"demand" form:"demand"`
-  Status int64 `json:"status" form:"status"`
+  Status int8 `json:"status" form:"status"`
 }
