+++ app/model/entity/user_footer.go
@@ -4,5 +4,5 @@ type UserFooter struct {
   Model
   Uid int64 `json:"uid" form:"uid"`
   Tuid int64 `json:"tuid" form:"tuid"`
-  IsRead int64 `json:"is_read" form:"is_read"`
+  IsRead int8 `json:"is_read" form:"is_read"`
 }
