+++ app/model/entity/user_reward.go
@@ -5,6 +5,6 @@ type UserReward struct {
   Uid int64 `json:"uid" form:"uid"`
   Tid int64 `json:"tid" form:"tid"`
   Tuid int64 `json:"tuid" form:"tuid"`
-  Gid int64 `json:"gid" form:"gid"`
+  Gid int32 `json:"gid" form:"gid"`
   Gamount float64 `json:"gamount" form:"gamount"`
 }
