+++ app/model/entity/wish_goods.go
@@ -10,9 +10,9 @@ type WishGoods struct {
   Adv string `json:"adv" form:"adv"`
   Domain string `json:"domain" form:"domain"`
   Img string `json:"img" form:"img"`
-  SendType int64 `json:"send_type" form:"send_type"`
-  Num int64 `json:"num" form:"num"`
-  SalerNum int64 `json:"saler_num" form:"saler_num"`
-  UserLimit int64 `json:"user_limit" form:"user_limit"`
+  SendType int8 `json:"send_type" form:"send_type"`
+  Num int32 `json:"num" form:"num"`
+  SalerNum int32 `json:"saler_num" form:"saler_num"`
+  UserLimit int32 `json:"user_limit" form:"user_limit"`
   PlanDate string `json:"plan_date" form:"plan_date"`
 }
