+++ app/model/entity/wish_support.go
@@ -8,22 +8,22 @@ type WishSupport struct {
   Uid int64 `json:"uid" form:"uid"`
   Gid int64 `json:"gid" form:"gid"`
   Gname string `json:"gname" form:"gname"`
-  Num int64 `json:"num" form:"num"`
+  Num int32 `json:"num" form:"num"`
   Price float64 `json:"price" form:"price"`
   Amount float64 `json:"amount" form:"amount"`
   CouponsId int64 `json:"coupons_id" form:"coupons_id"`
   NeedAmount float64 `json:"need_amount" form:"need_amount"`
   PayAmount float64 `json:"pay_amount" form:"pay_amount"`
-  PayType int64 `json:"pay_type" form:"pay_type"`
-  SendType int64 `json:"send_type" form:"send_type"`
+  PayType int8 `json:"pay_type" form:"pay_type"`
+  SendType int8 `json:"send_type" form:"send_type"`
   Virtual string `json:"virtual" form:"virtual"`
-  IsUse int64 `json:"is_use" form:"is_use"`
+  IsUse int8 `json:"is_use" form:"is_use"`
   Realname string `json:"realname" form:"realname"`
   Phone string `json:"phone" form:"phone"`
   Addr string `json:"addr" form:"addr"`
   Remark string `json:"remark" form:"remark"`
   SysRemark string `json:"sys_remark" form:"sys_remark"`
-  Status int64 `json:"status" form:"status"`
+  Status int8 `json:"status" form:"status"`
   ExpressName string `json:"express_name" form:"express_name"`
   ExpressCode string `json:"express_code" form:"express_code"`
   ExpressSn string `json:"express_sn" form:"express_sn"`
